@@ -1,6 +1,6 @@
 package com.polidea.multiplatformbleadapter;
 
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 import com.polidea.multiplatformbleadapter.utils.Constants.ConnectionPriority;
 
