@@ -1,6 +1,6 @@
 package com.polidea.multiplatformbleadapter;
 
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 import java.util.UUID;
 
