@@ -1,18 +1,18 @@
 package com.polidea.multiplatformbleadapter.utils;
 
 import android.bluetooth.BluetoothGatt;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
-import com.polidea.rxandroidble.RxBleCustomOperation;
-import com.polidea.rxandroidble.internal.RxBleLog;
-import com.polidea.rxandroidble.internal.connection.RxBleGattCallback;
+import com.polidea.rxandroidble2.RxBleCustomOperation;
+import com.polidea.rxandroidble2.internal.RxBleLog;
+import com.polidea.rxandroidble2.internal.connection.RxBleGattCallback;
 
 import java.lang.reflect.Method;
 import java.util.concurrent.Callable;
 import java.util.concurrent.TimeUnit;
 
-import rx.Observable;
-import rx.Scheduler;
+import io.reactivex.Observable;
+import io.reactivex.Scheduler;
 
 
 public class RefreshGattCustomOperation implements RxBleCustomOperation<Boolean> {
@@ -23,34 +23,31 @@ public class RefreshGattCustomOperation implements RxBleCustomOperation<Boolean>
             final BluetoothGatt bluetoothGatt,
             final RxBleGattCallback rxBleGattCallback,
             final Scheduler scheduler
-    ) throws Throwable {
+    ) {
 
-        return Observable.amb(
-                Observable.fromCallable(new Callable<Boolean>() {
-                    @Override
-                    public Boolean call() throws Exception {
-                        boolean success = false;
-                        call: try {
-                            Method bluetoothGattRefreshFunction = bluetoothGatt.getClass().getMethod("refresh");
-                            if (bluetoothGattRefreshFunction == null) {
-                                RxBleLog.d("Could not find function BluetoothGatt.refresh()");
-                                break call;
-                            }
-
-                            success = (Boolean) bluetoothGattRefreshFunction.invoke(bluetoothGatt);
-
-                            if (!success) RxBleLog.d("BluetoothGatt.refresh() returned false");
-                        } catch (Exception e) {
-                            RxBleLog.d(e, "Could not call function BluetoothGatt.refresh()");
+        return Observable.ambArray(
+                Observable.fromCallable(() -> {
+                    boolean success = false;
+                    call:
+                    try {
+                        Method bluetoothGattRefreshFunction = bluetoothGatt.getClass().getMethod("refresh");
+                        if (bluetoothGattRefreshFunction == null) {
+                            RxBleLog.d("Could not find function BluetoothGatt.refresh()");
+                            break call;
                         }
 
-                        RxBleLog.i("Calling BluetoothGatt.refresh() status: %s", success ? "Success" : "Failure");
-                        return success;
+                        success = (Boolean) bluetoothGattRefreshFunction.invoke(bluetoothGatt);
+                        if (!success) RxBleLog.d("BluetoothGatt.refresh() returned false");
+                    } catch (Exception e) {
+                        RxBleLog.d(e, "Could not call function BluetoothGatt.refresh()");
                     }
+
+                    RxBleLog.i("Calling BluetoothGatt.refresh() status: %s", success ? "Success" : "Failure");
+                    return success;
                 })
                         .subscribeOn(scheduler)
                         .delay(1, TimeUnit.SECONDS, scheduler),
-                rxBleGattCallback.<Boolean>observeDisconnect()
+                rxBleGattCallback.observeDisconnect()
         );
     }
 }
